Rename handleNetworkByQR to EIP618Helper
